requirements.txt
python-dotenv==1.0.1
llama-parse==0.4.0
langchain==0.2.14
langchain-community==0.2.12
langchain-huggingface==0.0.3
sentence-transformers==3.0.1
chromadb>=0.5.0,<0.5.4
langchain-chroma
langchain-groq==0.1.9
streamlit==1.37.1